requirements.txt
dj-database-url==0.4.0
Django==1.9.2
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
gunicorn==19.4.5
mysql==0.0.1
MySQL-python==1.2.5
Pillow==3.1.1
psycopg2==2.6.1
requests==2.9.1
slugify==0.0.1
wheel==0.24.0
whitenoise==2.0.6
